backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.9.1
anthropic==0.39.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
beautifulsoup4==4.12.3
bcrypt==4.0.1
pytest==8.3.4
pytest-httpx==0.35.0
